♻️ 1F9GZD task: keep evaluator review assembly below hotspot cap
