+++ .agentplane/tasks/202607291148-1F9GZD/pr/meta.json
@@ -2,7 +2,7 @@
   "base": "main",
   "branch": "task/202607291148-1F9GZD/formalize-sha-bound-qualification-packets-for-ev",
   "created_at": "2026-07-29T11:49:50.615Z",
-  "diffstat_sha256": "sha256:d33b8a3e2b41a7ba6f16c9cdb0e5f89cf325e622d8f36186044443d7469c4040",
+  "diffstat_sha256": "sha256:2e9dc7fdf28c7368ada6cd271cdd600eaae9d4b3b0e462958443cfc10f8c278a",
   "last_verified_at": "2026-07-29T14:19:04.724Z",
   "last_verified_diffstat_sha256": "sha256:d33b8a3e2b41a7ba6f16c9cdb0e5f89cf325e622d8f36186044443d7469c4040",
   "pr_number": 4670,
