+++ .agentplane/tasks/202607291148-1F9GZD/pr/review.md
@@ -32,7 +32,8 @@ Created: 2026-07-29T11:49:50.615Z
  .../evaluator-qualification-packet.test.ts         | 581 +++++++++++++++++++++
  .../evaluator/evaluator-qualification-review.ts    |  52 ++
  .../evaluator/evaluator-review-artifacts.ts        |   3 +-
- .../commands/evaluator/evaluator-review-usecase.ts |  47 +-
+ .../commands/evaluator/evaluator-review-shared.ts  |  20 +
+ .../commands/evaluator/evaluator-review-usecase.ts |  76 +--
  .../evaluator/evaluator-run.command.test.ts        |  72 +--
  .../commands/evaluator/evaluator-test-helpers.ts   |  72 +++
  .../evaluator/evaluator-verification-records.ts    |  17 +-
@@ -44,7 +45,7 @@ Created: 2026-07-29T11:49:50.615Z
  .../src/commands/task/qualification-packet-rf04.ts | 256 +++++++++
  .../src/commands/task/qualification-packet.ts      | 556 ++++++++++++++++++++
  .../src/commands/task/verify-record-execute.ts     |  33 +-
- 15 files changed, 1881 insertions(+), 105 deletions(-)
+ 16 files changed, 1909 insertions(+), 126 deletions(-)
 ```
 
 </details>
